requirements.txt
# annotated-types==0.6.0
# anyio==4.3.0
# APScheduler==3.10.4
# asttokens==2.4.1
# attrs==24.2.0
# backcall==0.2.0
# beautifulsoup4==4.12.3
# bleach==6.1.0
# boto3==1.34.147
# botocore==1.34.147
# certifi==2024.2.2
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# decorator==5.1.1
# defusedxml==0.7.1
# dnspython==2.6.1
# docopt==0.6.2
email_validator==2.1.1
# executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
fastjsonschema==2.20.0
# geopandas==1.0.1
# h11==0.14.0
# httpcore==1.0.5
# httptools==0.6.1
# httpx==0.27.0
# idna==3.7
# ipython==8.12.3
# itsdangerous==2.2.0
# jedi==0.19.1
# Jinja2==3.1.4
# jmespath==1.0.1
# jsonschema==4.23.0
# jsonschema-specifications==2023.12.1
# jupyter_client==8.6.2
# jupyter_core==5.7.2
# jupyterlab_pygments==0.3.0
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# mistune==3.0.2
mysql-connector-python==8.4.0
# nbclient==0.10.0
# nbconvert==7.16.4
# nbformat==5.10.4
# numpy==2.0.1
# orjson==3.10.3
# packaging==24.1
# pandas==2.2.2
# pandocfilters==1.5.1
# parso==0.8.4
phonenumbers==8.13.40
# pickleshare==0.7.5
# pillow==10.4.0
# pipreqs==0.5.0
# platformdirs==4.3.3
# prompt_toolkit==3.0.47
# pure_eval==0.2.3
pydantic==2.7.1
pydantic-extra-types==2.8.2
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
# pyogrio==0.9.0
# pyproj==3.6.1
# python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# python-multipart==0.0.9
# pytz==2024.1
# pywin32==306
# PyYAML==6.0.1
# pyzmq==26.2.0
# redis==5.0.8
# referencing==0.35.1
requests==2.32.3
# rich==13.7.1
# rpds-py==0.20.0
s3transfer==0.10.2
# shapely==2.0.5
# shellingham==1.5.4
# six==1.16.0
# sniffio==1.3.1
# soupsieve==2.6
# stack-data==0.6.3
starlette==0.37.2
# tinycss2==1.3.0
# tornado==6.4.1
# traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
# tzdata==2024.1
# tzlocal==5.2
# ujson==5.9.0
urllib3==2.2.2
uvicorn==0.29.0
# watchfiles==0.21.0
# wcwidth==0.2.13
# webencodings==0.5.1
# websockets==12.0
# yarg==0.1.9
